rename module interpreter to context
