+++ src/learning_lab/01_device_control.py
@@ -23,7 +23,7 @@
 from __future__ import print_function as _print_function
 from pydoc import plain
 from pydoc import render_doc as doc
-from basics.interpreter import sys_exit, EX_OK, EX_TEMPFAIL
+from basics.context import sys_exit, EX_OK, EX_TEMPFAIL
 from basics.inventory import inventory_mounted, device_control, DeviceControl
 from basics.render import print_table
