+++ src/learning_lab/01_device_dismount.py
@@ -20,7 +20,7 @@ from __future__ import print_function as _print_function
 from basics.inventory import device_dismount, inventory_mounted
 from pydoc import render_doc as doc
 from pydoc import plain
-from basics.interpreter import sys_exit, EX_OK, EX_TEMPFAIL
+from basics.context import sys_exit, EX_OK, EX_TEMPFAIL
 
 def demonstrate(device_name):
     print('device_dismount(' + device_name, end=')\n')
