+++ src/learning_lab/03_interface_shutdown.py
@@ -23,7 +23,7 @@
 from __future__ import print_function as _print_function
 from pydoc import plain
 from pydoc import render_doc as doc
-from basics.interpreter import sys_exit, EX_OK, EX_TEMPFAIL
+from basics.context import sys_exit, EX_OK, EX_TEMPFAIL
 from basics.interface import management_interface
 from basics.interface import interface_configuration_tuple
 from basics.interface import interface_configuration_update
