+++ src/learning_lab/04_static_route_suite.py
@@ -38,7 +38,7 @@
         04_static_route_suite (this file)
 '''
 
-from basics.interpreter import EX_OK
+from basics.context import EX_OK
 from runpy import run_module
 
 def run_script(script):
