+++ src/learning_lab/05_acl_capability.py
@@ -21,7 +21,7 @@ If there are no such devices then all sample scripts prefixed with
 
 from __future__ import print_function
 from pydoc import render_doc as doc, plain
-from basics.interpreter import sys_exit, EX_OK, EX_TEMPFAIL
+from basics.context import sys_exit, EX_OK, EX_TEMPFAIL
 from basics.acl import capability_ns, capability_name
 from basics.inventory import capability_discovery
 from basics.render import print_table
