+++ src/learning_lab/05_acl_exists.py
@@ -21,7 +21,7 @@
 from __future__ import print_function as _print_function
 from pydoc import plain
 from pydoc import render_doc as doc
-from basics.interpreter import sys_exit, EX_OK, EX_TEMPFAIL
+from basics.context import sys_exit, EX_OK, EX_TEMPFAIL
 from basics.acl import acl_exists, inventory_acl
 from importlib import import_module
 acl_fixture = import_module('learning_lab.05_acl_fixture')
