+++ src/learning_lab/05_acl_list.py
@@ -22,7 +22,7 @@
 from __future__ import print_function as _print_function
 from pydoc import plain
 from pydoc import render_doc as doc
-from basics.interpreter import sys_exit, EX_OK, EX_TEMPFAIL
+from basics.context import sys_exit, EX_OK, EX_TEMPFAIL
 from basics.acl import acl_list,inventory_acl
 
 def demonstrate(device_name):
