+++ src/learning_lab/restore_network_state.py
@@ -40,7 +40,7 @@
     * Devices not configured in the settings
 """
 
-from basics.interpreter import EX_OK
+from basics.context import EX_OK
 from runpy import run_module
 
 def run_script(script):
